+++ CHANGELOG.md
@@ -6,9 +6,11 @@ All notable changes to eww will be listed here, starting at changes since versio
 
 ### Fixes
 - Fix and refactor nix flake (By: w-lfchen)
+- Fix remove items from systray (By: vnva)
 
 ### Features
 - Add `:truncate` property to labels, disabled by default (except in cases where truncation would be enabled in version `0.5.0` and before) (By: Rayzeq).
+- Add support for `:hover` css selectors for tray items (By: zeapoz)
 
 ## [0.6.0] (21.04.2024)